feat: searchoverlay 수정구현 (#19)
